May now specify a cross-linker name override.
This converter contains a map between standard plink 2 cross-linker
names and cross-linker name supported by proxl. However, if a custom
definition is made for one of these linkers (e.g., a custom mono-link
mass for EDC), this new linker definition will be given a name that the
converter doesn't recognize. This will cause an error, since it can't
translate to a proxl cross-linker name.

Users may now specify -l linker_name to specify which proxl-supported
cross-linker name should be used for this cross-linker. This name will
always be used for the linker.